+++ packages/terra-form-select/tests/wdio/_frame-spec.js
@@ -1,0 +1,14 @@
+/* global before, browser, Terra */
+const viewports = Terra.viewports('tiny');
+
+describe('Frame', () => {
+  describe("should have a themeable disabled icon", () => {
+    before(() => browser.url('/#/raw/tests/terra-form-select/form-select/disabled'));
+
+    Terra.should.beAccessible({ viewports });
+    Terra.should.matchScreenshot({ viewports });
+    Terra.should.themeEachCustomProperty({
+      '--terra-select-disabled-arrow-background': 'rgb(0, 0, 255)',
+    });
+  });
+});